refactor: 아바타 레벨 관련 코드 AvatarService 생성하여 이동
